refactor to combine mirations and postgres into a single library
