+++ libraries/golang/postgres/go.mod
@@ -1,0 +1,9 @@
+module ponglehub.co.uk/auth/postgres
+
+go 1.16
+
+require (
+	github.com/jackc/pgx/v4 v4.13.0
+	github.com/sirupsen/logrus v1.8.1
+	github.com/stretchr/testify v1.7.0
+)
