+++ libraries/golang/postgres/pkg/types/types.go
@@ -1,0 +1,15 @@
+package types
+
+import (
+	"ponglehub.co.uk/auth/postgres/pkg/connect"
+)
+
+type Migration struct {
+	Query string
+}
+
+type MigrationConfig struct {
+	AdminConfig  connect.ConnectConfig
+	TargetConfig connect.ConnectConfig
+	Migrations   []Migration
+}
