+++ services/auth-server/cmd/users-get/users-get.go
@@ -4,11 +4,10 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 	"github.com/sirupsen/logrus"
-	"ponglehub.co.uk/auth/auth-server/internal/client"
 	"ponglehub.co.uk/auth/auth-server/internal/server"
 )
 
-func RouteBuilder(cli client.AuthClient, r *gin.Engine) {
+func RouteBuilder(cli server.AuthClient, r *gin.Engine) {
 	r.GET("/:id", func(c *gin.Context) {
 		id := c.Param("id")
 		if _, err := uuid.Parse(id); err != nil {
