--- services/db-init/go.mod
@@ -1,10 +1,0 @@
-module ponglehub.co.uk/auth/db-init
-
-go 1.16
-
-require (
-	github.com/jackc/pgx/v4 v4.11.0
-	github.com/sirupsen/logrus v1.8.1
-	github.com/stretchr/testify v1.5.1
-	gopkg.in/yaml.v2 v2.4.0 // indirect
-)
